create test data generation class
